Join connection and database containers
There is not need to keep them separated
