+++ src/renderer/entry.jsx
@@ -6,9 +6,7 @@ import { Router, Route } from 'react-router';
 import App from './containers/app.jsx';
 import configureStore from './store/configure';
 import ServerManagementContainer from './containers/server-management.jsx';
-// import QueryBrowserContainer from './containers/query-browser.jsx';
-import ConnectionContainer from './containers/connection.jsx';
-import DatabaseContainer from './containers/database.jsx';
+import QueryBrowserContainer from './containers/query-browser.jsx';
 
 
 const history = createHashHistory();
@@ -20,10 +18,7 @@ ReactDOM.render(
     <Router history={history}>
       <Route component={App}>
         <Route path="/" component={ServerManagementContainer} />
-        {/* <Route path="/server/:id" component={QueryBrowserContainer} />*/}
-        <Route component={ConnectionContainer}>
-          <Route path="/server/:id/database/:database" component={DatabaseContainer} />
-        </Route>
+        <Route path="/server/:id/database/:database" component={QueryBrowserContainer} />
       </Route>
     </Router>
   </Provider>,
